Switch back to actual webpack.config
Didn't know it supports multi-config
