+++ scripts/sizeSnapshot/create.js
@@ -1,12 +1,10 @@
 const fse = require('fs-extra');
-const globCallback = require('glob');
 const lodash = require('lodash');
 const path = require('path');
 const { promisify } = require('util');
 const webpackCallbackBased = require('webpack');
 const createWebpackConfig = require('./webpack.config');
 
-const glob = promisify(globCallback);
 const webpack = promisify(webpackCallbackBased);
 
 const workspaceRoot = path.join(__dirname, '../../');
@@ -30,114 +28,13 @@ async function getRollupSize(snapshotPath) {
   ]);
 }
 
-async function getWebpackEntries() {
-  const corePackagePath = path.join(workspaceRoot, 'packages/material-ui/build');
-  const coreComponents = (await glob(path.join(corePackagePath, '[A-Z]*/index.js'))).map(
-    (componentPath) => {
-      const componentName = path.basename(path.dirname(componentPath));
-      let entryName = componentName;
-      // adjust for legacy names
-      if (componentName === 'Paper') {
-        entryName = '@material-ui/core/Paper.esm';
-      } else if (componentName === 'TextareaAutosize') {
-        entryName = '@material-ui/core/Textarea';
-      } else if (['Popper'].indexOf(componentName) !== -1) {
-        entryName = `@material-ui/core/${componentName}`;
-      }
-
-      return {
-        name: entryName,
-        path: path.relative(workspaceRoot, path.dirname(componentPath)),
-      };
-    },
-  );
-
-  const labPackagePath = path.join(workspaceRoot, 'packages/material-ui-lab/build');
-  const labComponents = (await glob(path.join(labPackagePath, '[A-Z]*/index.js'))).map(
-    (componentPath) => {
-      const componentName = path.basename(path.dirname(componentPath));
-
-      return {
-        name: componentName,
-        path: path.relative(workspaceRoot, path.dirname(componentPath)),
-      };
-    },
-  );
-
-  return [
-    {
-      name: '@material-ui/core',
-      path: path.join(path.relative(workspaceRoot, corePackagePath), 'index.js'),
-    },
-    {
-      name: '@material-ui/lab',
-      path: path.join(path.relative(workspaceRoot, labPackagePath), 'index.js'),
-    },
-    {
-      name: '@material-ui/styles',
-      path: 'packages/material-ui-styles/build/index.js',
-    },
-    {
-      name: '@material-ui/system',
-      path: 'packages/material-ui-system/build/esm/index.js',
-    },
-    ...coreComponents,
-    {
-      name: '@material-ui/core/styles/createMuiTheme',
-      path: 'packages/material-ui/build/styles/createMuiTheme.js',
-    },
-    {
-      name: 'colorManipulator',
-      path: 'packages/material-ui/build/styles/colorManipulator.js',
-    },
-    ...labComponents,
-    {
-      name: 'useAutocomplete',
-      path: 'packages/material-ui-lab/build/useAutocomplete/index.js',
-    },
-    {
-      name: '@material-ui/core/useMediaQuery',
-      path: 'packages/material-ui/build/useMediaQuery/index.js',
-    },
-    {
-      name: '@material-ui/core/useScrollTrigger',
-      path: 'packages/material-ui/build/useScrollTrigger/index.js',
-    },
-    {
-      name: '@material-ui/utils',
-      path: 'packages/material-ui-utils/build/esm/index.js',
-    },
-    // TODO: Requires webpack v5
-    // Resolution of webpack/acorn to 7.x is blocked by nextjs (https://github.com/vercel/next.js/issues/11947)
-    // {
-    //   name: '@material-ui/core.modern',
-    //   webpack: true,
-    //   path: path.join(path.relative(workspaceRoot, corePackagePath), 'modern/index.js'),
-    // },
-    {
-      name: '@material-ui/core.legacy',
-      path: path.join(path.relative(workspaceRoot, corePackagePath), 'legacy/index.js'),
-    },
-  ];
-}
-
 /**
  * creates size snapshot for every bundle that built with webpack
  */
 async function getWebpackSizes() {
   await fse.mkdirp(path.join(__dirname, 'build'));
-  const [baseConfig, entries] = await Promise.all([
-    createWebpackConfig(webpack),
-    getWebpackEntries(),
-  ]);
 
-  const configurations = entries.map((entry) => {
-    return {
-      ...baseConfig,
-      entry: { [entry.name]: path.join(workspaceRoot, entry.path) },
-    };
-  });
-
+  const configurations = await createWebpackConfig(webpack);
   const webpackMultiStats = await webpack(configurations);
 
   const sizes = [];
